make the document more generic

+++ src/indexes/mod.rs
@@ -7,7 +7,7 @@ pub trait Index: Send + Sync {
     fn get_documents(&self) -> Vec<Document>;
 
     /// Get one document in the index
-    fn get_document(&self, id: usize) -> Option<Document>;
+    fn get_document(&self, id: u32) -> Option<Document>;
 
     /// Add a batch of documents
     fn add_documents(&mut self, document: Vec<Document>);
